+++ tests/Jroc.Test262.Tests/built-ins/String/prototype/localeCompare/ExecutionTests.CanonicalEquivalence.cs
@@ -1,0 +1,7 @@
+namespace Jroc.Test262.Tests.built_ins.String.prototype.localeCompare;
+
+public partial class ExecutionTests
+{
+    [Fact(DisplayName = "15.5.4.9_CE.js")]
+    public Task _15_5_4_9_CE() => ExecutionTestFromFile("15.5.4.9_CE");
+}