+++ tests/Jroc.Test262.Tests/built-ins/String/prototype/toLocaleLowerCase/ExecutionTests.UnicodeCasing.cs
@@ -1,0 +1,10 @@
+namespace Jroc.Test262.Tests.built_ins.String.prototype.toLocaleLowerCase;
+
+public partial class ExecutionTests
+{
+    [Fact(DisplayName = "Final_Sigma_U180E.js")]
+    public Task Final_Sigma_U180E() => ExecutionTestFromFile("Final_Sigma_U180E");
+
+    [Fact(DisplayName = "special_casing_conditional.js")]
+    public Task special_casing_conditional() => ExecutionTestFromFile("special_casing_conditional");
+}